agua_molino_3/django_project/requirements.txt
alabaster==0.7.12
anaconda-client==1.11.0
anaconda-navigator==2.0.4
anaconda-project==0.11.1
anyio==3.5.0
appdirs==1.4.4
argh==0.26.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
asn1crypto==1.5.1
astroid==2.14.2
astropy==4.2.1
asttokens==2.0.5
async-generator==1.10
atomicwrites==1.4.0
attrs==22.1.0
autopep==81.5.6
Babel==2.11.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
backports.zoneinfo==0.2.1
bcrypt==3.2.0
beautifulsoup==44.11.1
bitarray==2.5.1
bkcharts==0.2
black==22.6.0
bleach==4.1.0
bokeh==2.3.3
boto==2.49.0
Bottleneck==1.3.5
bpyutils==0.5.8
brotlipy==0.7.0
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.4
cloudpickle==2.0.0
clyent==1.2.2
colorama==0.4.6
comm==0.1.2
comtypes==1.1.10
conda==23.1.0
conda-build==3.21.4
conda-content-trust==0.1.3
conda-pack==0.6.0
conda-package-handling==2.0.2
conda_package_streaming==0.7.0
conda-repo-cli==1.0.27
conda-token==0.4.0
conda-verify==3.4.2
contextlib==221.6.0
crispy-forms-gds==0.2.4
cryptography==38.0.4
cycler==0.11.0
Cython==0.29.33
cytoolz==0.12.0
dask==2021.7.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
dill==0.3.6
distributed==2021.7.0
Django==3.2.5
django-crispy-forms==1.12.0
django-embed-video==1.4.0
django-environ==0.4.5
django-js-routes==0.1.4
django-password-validators==1.6.1
django-proxy-server==0.6
django-rest-framework==0.1.0
django-revproxy==0.10.0
django-session-timeout==0.1.0
Django-Verify-Email==1.0.7
djangorestframework==3.12.4
docutils==0.18.1
entrypoints==0.4
et-xmlfile==1.1.0
excel2json==1.0.1
excel2json-3==0.1.6
executing==0.8.3
fastcache==1.1.0
fastjsonschema==2.16.2
filelock==3.9.0
flake==83.9.0
Flask==2.2.2
flit_core==3.6.0
fsspec==2022.11.0
future==0.18.2
gevent==22.10.2
glob==20.7
gmpy==22.1.2
greenlet==2.0.1
h5py==3.7.0
HeapDict==1.0.1
html5lib==1.1
idna==3.4
imagecodecs==2021.6.8
imageio==2.9.0
imagesize==1.4.1
importlib-metadata==4.11.3
importlib-resources==5.2.0
iniconfig==1.1.1
intervaltree==3.1.0
ipykernel==6.19.2
ipython==8.10.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==5.9.3
itsdangerous==2.0.1
jdcal==1.4.1
jedi==0.17.2
Jinja==23.1.2
joblib==1.1.1
json==50.9.6
jsonschema==4.17.3
jupyter==1.0.0
jupyter_client==7.4.9
jupyter-console==6.4.4
jupyter_core==5.2.0
jupyter-packaging==0.7.12
jupyter-server==1.23.4
jupyterlab==3.5.3
jupyterlab-pygments=0.1.2
jupyterlab_server==2.19.0
jupyterlab-widgets==1.0.0
keyring==23.4.0
kiwisolver==1.4.4
lazy-object-proxy==1.6.0
libarchive-c==2.9
llvmlite==0.39.1
locket==1.0.0
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.3.4
matplotlib-inline==0.1.6
mccabe==0.6.1
menuinst==1.4.16
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mock==4.0.3
more-itertools==8.12.0
mpmath==1.2.1
msgpack==1.0.3
multipledispatch==0.6.0
mypy-extensions==0.4.3
navigator-updater==0.2.1
nbclassic==0.5.2
nbclient==0.5.13
nbconvert==6.5.4
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.8.4
nltk==3.7
nose==1.3.7
notebook==6.5.2
notebook_shim==0.2.2
numba==0.56.4
numexpr==2.8.4
numpy==1.23.5
numpydoc==1.5.0
olefile==0.46
openpyxl==3.0.10
packaging==22.0
pandas==1.5.2
pandocfilters==1.5.0
paramiko==2.8.1
parso==0.7.0
partd==1.2.0
path==16.2.0
pathlib==1.0.1
pathlib==22.3.6
pathspec==0.10.3
pathtools==0.1.2
patsy==0.5.3
pep==81.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==22.3.1
pkginfo==1.8.3
pkgutil_resolve_name==1.3.10
platformdirs==2.5.2
pluggy==1.0.0
ply==3.11
pooch==1.4.0
pproxy==2.7.8
prometheus-client==0.14.1
prompt-toolkit==3.0.36
psutil==5.9.0
psycopg==22.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pvpn==0.2.5
py==1.11.0
pyasn==10.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycosat==0.6.4
pycparser==2.21
pycryptodome==3.10.1
pycurl==7.45.1
pydocstyle==6.3.0
pyerfa==2.0.0
pyflakes==2.2.0
Pygments==2.11.2
PyJWT==2.6.0
pylint==2.16.2
pyls-black==0.4.6
pyls-spyder==0.3.2
PyNaCl==1.5.0
pyodbc==4.0.34
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyreadline==2.1
pyrsistent==0.18.0
PySocks==1.7.1
pytest==7.1.2
python-dateutil==2.8.2
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
pytoolconfig==1.2.5
pytz==2022.7.1
PyWavelets==1.4.1
pywin==32305.1
pywin32-ctypes==0.2.0
pywinpty==2.0.2
PyYAML==6.0
pyzmq==23.2.0
QDarkStyle==2.8.1
QtAwesome==1.2.2
qtconsole==5.4.0
QtPy==2.2.0
regex==2022.7.9
requests==2.28.1
rope==1.7.0
rsa==4.9
Rtree==1.0.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
ruamel-yaml-conda==0.17.21
scikit-image==0.18.1
scikit-learn==1.2.1
scipy==1.10.0
seaborn==0.11.1
Send2Trash==1.8.0
setuptools==65.6.3
simplegeneric==0.8.1
singledispatch==3.7.0
sip==4.19.13
six==1.16.0
sniffio==1.2.0
snowballstemmer==2.2.0
sortedcollections==2.1.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
Sphinx==5.0.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sphinxcontrib-websupport==1.2.4
spyder==4.2.5
spyder-kernels==1.10.2
SQLAlchemy==1.4.39
sqlparse==0.4.3
stack-data==0.2.0
statsmodels==0.13.5
style==1.1.0
sympy==1.11.1
tables==3.7.0
tblib==1.7.0
terminado==0.17.1
testpath==0.6.0
textdistance==4.2.1
threadpoolctl==2.2.0
three-merge==0.1.1
tifffile==2021.7.2
tinycss==21.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.1
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.7.1
typed-ast==1.4.3
typing_extensions==4.4.0
tzdata==2021.1
ujson==5.4.0
unicodecsv==0.14.1
Unipath==1.1
update==0.0.1
uritemplate==4.1.1
urllib==31.26.14
watchdog==1.0.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
Werkzeug==2.2.2
wfastcgi==3.0.0
wheel==0.38.4
widgetsnbextension==3.5.2
win-inet-pton==1.1.0
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.14.1
xlrd==2.0.1
XlsxWriter==3.0.3
xlwings==0.29.1
xlwt==1.3.0
xmltodict==0.12.0
yapf==0.31.0
zict==2.1.0
zipp==3.11.0
zope.event==4.5.0
zope.interface==5.4.0
zstandard==0.19.0
